==> course/Python - SGBDS/07. Google BigQuery/4.Carregando Dados para o Google BigQuery.ipynb <==
#pip install google-cloud-bigquery
#pip install --upgrade google-cloud-bigquery
#pip install pandas

#pip install pandas pandas-gbq

import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq

# Carregue suas credenciais
credencial = service_account.Credentials.from_service_account_file(
    r'C:\Users\edmil\Downloads\pythoncurso-398511-260383f13f39.json',
    scopes=['https://www.googleapis.com/auth/bigquery']
)

df = pd.read_excel(r"C:\Users\edmil\OneDrive\Área de Trabalho\Material Python-Google BigQuery\Arquivos\Origem\arquivos_excel\Produto.xlsx")
df.head()

# Envie o DataFrame para o BigQuery
df.to_gbq(destination_table='pythoncurso-398511.AulaPython.Produto',  # Substitua pelo nome correto da tabela(obs caminho completo)
          project_id='pythoncurso-398511',
          if_exists='replace',
          credentials=credencial)

#replace = subistuir a antiga pela original 
#append = adiciona dados no fim da tabela
