==> course/Python - SGBDS/07. Google BigQuery/9.Mantendo histórico de base.ipynb <==
#### Trabalhando com Datas e mantendo histórico

import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq
from datetime import datetime

df = pd.read_excel(r"K:\Python\Arquivos\Origem\arquivos_excel\Produto.xlsx")
df.head()

df['DataHoraCarga'] = datetime.now()

df.head(5)

df['DataHoraCarga'] = datetime.now().strftime('%Y-%m-%d %H:%M:%S')  # prescisa desta conversao para o bigquery aceitar 

df.head(5)


df['DataCarga'] = datetime.now().date()

df.head(5)

df['DataCargaPTBR'] = datetime.now().strftime('%d/%m/%Y')

df.head(5)

df['HoraCarga'] = datetime.now().strftime('%H:%M:%S')

df.head(5)

df.dtypes

##### https://docs.python.org/pt-br/3/library/datetime.html

#converter para BQ Aceitar
df['DataHoraCarga'] = pd.to_datetime(df['DataHoraCarga'])
df['DataCarga'] = pd.to_datetime(df['DataCarga'])

df.dtypes

df.head(5)

# credenciais para validar importar dados para o google bigquery
credencial = r"C:\Users\edmil\Downloads\pythoncurso-398511-260383f13f39.json"
CredImportacao = service_account.Credentials.from_service_account_file(
    credencial,
    scopes=['https://www.googleapis.com/auth/bigquery']
)

# variaveis do projeto 
projeto = 'pythoncurso-398511'
dataset = 'AulaPython'
tabela = 'ProdutoHistorico'
parameretro ='append'      
#replace = subistuir a antiga pela original 
#append = adiciona dados no fim da tabela 

#df['Price'] = df['Price'].astype(int)
#df['Id_Category'] = df['Id_Category'].astype(int)

df.dtypes

schema = [
    {'name': 'DataHoraCarga', 'type': 'TIMESTAMP'},# Use 'TIMESTAMP' para datetime64[ns] (pode ser tambem da certo)
    {'name': 'Price', 'type': 'INTEGER'},
    {'name': 'Id_Category', 'type': 'INTEGER'},
    {'name': 'DataCarga', 'type': 'DATE'} #SOMENTE DATA
]



# Use o método to_gbq para enviar o DataFrame para o BigQuery
df.to_gbq(destination_table=f'{projeto}.{dataset}.{tabela}',
          project_id=projeto,
          if_exists=parameretro,
          credentials=CredImportacao,
          table_schema=schema)

"""

SELECT * FROM `pythoncurso-398511.AulaPython.ProdutoHistorico`
where HoraCarga = 
  (SELECT max(HoraCarga)FROM `pythoncurso-398511.AulaPython.ProdutoHistorico`)

"""
