==> course/Python - SGBDS/07. Google BigQuery/11.Conexão com Google Sheets.ipynb <==
#pip install gspread 
#pip install oauth2client

import pandas as pd
import gspread
from oauth2client.service_account import ServiceAccountCredentials
# Video referencia para criar Credenciao (Configurando Credencias de Acesso ao Google BigQuery)
# Caminho para o arquivo JSON de credenciais

credencialjson = r"C:\Users\edmil\Downloads\pythoncurso-398511-260383f13f39.json"

# Configuração das credenciais
scope = ["https://spreadsheets.google.com/feeds", "https://www.googleapis.com/auth/drive"]

credentials = ServiceAccountCredentials.from_json_keyfile_name(credencialjson, scope)

# Autentique-se com o Google Sheets
gc = gspread.authorize(credentials)

# Abra o arquivo do Google Sheets pelo seu ID 
idplanilha = gc.open_by_key("1Dfem4ThXXdfgAW9aA6G0jJES1yUsjXln24DPaxt8JE8")

# Acesse a guia da planilha
guiaplanilha = idplanilha.worksheet("Produto")

# ler dados
Dados = guiaplanilha.get_all_values()

# Exiba os dados 
"""
for row in Dados:
    print(row)
    
"""


Produto = pd.DataFrame(guiaplanilha.get_all_records())
Produto.head(10)

# Converter os dados em um DataFrame
#df = pd.DataFrame(Dados)
#df.head(10)

dftratado = pd.DataFrame(Dados[1:], columns=Dados[0])
dftratado.head()

#### Apis Google
https://cloud.google.com/docs/authentication/api-keys?hl=pt-br
